+++ R/bicop.R
@@ -307,10 +307,11 @@ bicop_dist <- function(
       parameters <- as.matrix(parameters)
       stopifnot(dim(parameters) == c(30, 30))
       margin_integrals <- c(rowMeans(parameters), colMeans(parameters))
-      if (any(abs(margin_integrals - 1) > 0.5))
+      if (any(abs(margin_integrals - 1) > 0.5)) {
         warning(
           "margins implied by 'parameters' deviate strongly from the standard uniform distribution."
         )
+      }
     } else {
       parameters <- matrix(1, 30, 30)
     }